// rules/interfaces.ts
export interface BAPEnterprice {
    id: number;
    shortName: string;
}
export interface BAPRole {
    id: number;
    name: string;
    user: string;
}

export interface bodyPerson {
    name: string;
    lastname: string;
    email: string | null;
    password: string;
    phoneNumber: string | null;
    employeeNumber: string;
    enterprice: BAPEnterprice;
    role: BAPRole;
}
// ...
export interface optionsUpdatePerson {
    resetPassword?: boolean;
    deletePerson?: boolean;
    updateStatus?: 'ACTIVO' | 'INACTIVO';
    updateData?: bodyPerson;
}

export interface optionsUpdateEnterprice {
    deleteEnterprice?: boolean;
    updateData?: { shortName: string, name: string };
}

export interface actionPersonProps {
    person: ExistPerson;
    option: optionsUpdatePerson
}

export interface actionEnterpriceProps {
    enterprice: BAPEnterprice;
    option: optionsUpdateEnterprice;
}
// ...
export interface ExistPerson { id: string, role: number, name: string };
// ...
export interface user {
    codigo: string | null;
    nombre: string | null;
    clave: string | null;
    descripcion: string | null;
}